requirements.txt
accelerate==1.3.0
av==12.0.0
diffusers==0.32.2
einops==0.8.0
matplotlib==3.10.1
moviepy==1.0.3
numpy==2.2.3
omegaconf==2.3.0
pillow==11.1.0
pytorch-fid==0.3.0
safetensors==0.5.2
tensorboard==2.18.0
timm==1.0.15
torchode==1.0.0
tqdm==4.67.1
gdown==5.2.0
--find-links https://download.pytorch.org/whl/torch/
torch==2.5.1+cu121
--find-links https://download.pytorch.org/whl/torchvision/
torchvision==0.20.1+cu121
--find-links https://download.pytorch.org/whl/torchaudio/
torchaudio==2.5.1+cu121

